requirements.txt
# PARTH AI System - Python Dependencies

# Core AI/ML Libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.3.0

# Multi-Agent Frameworks
autogen-agentchat>=0.2.0
crewai>=0.1.0
metagpt>=0.1.0

# Voice & Audio
openai-whisper>=20231117
TTS>=0.19.0
pydub>=0.25.1
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.13
librosa>=0.10.0

# Vision & Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
diffusers>=0.24.0
invisible-watermark>=0.2.0

# Web & API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
requests>=2.31.0
aiohttp>=3.9.0
gradio>=4.0.0

# Database & Storage
chromadb>=0.4.18
faiss-cpu>=1.7.4
sqlite3
redis>=5.0.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
tqdm>=4.66.1
psutil>=5.9.6
pydantic>=2.5.0
tenacity>=8.2.3

# Device Control (Android/Termux)
androidhelper>=0.1; sys_platform == 'linux'

# Platform-specific
pywin32>=306; sys_platform == 'win32'
pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'

# Development Tools
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional but recommended
jupyter>=1.0.0
ipython>=8.17.0
